Move artwork rail; wire up link
