+++ desktop/apps/auction/components/artwork_rail/ArtworkRail.js
@@ -1,5 +1,5 @@
-import ChevronLeft from '../../../../../components/main_layout/public/icons/chevron-left.svg'
-import ChevronRight from '../../../../../components/main_layout/public/icons/chevron-right.svg'
+import ChevronLeft from '../../../../components/main_layout/public/icons/chevron-left.svg'
+import ChevronRight from '../../../../components/main_layout/public/icons/chevron-right.svg'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import block from 'bem-cn-lite'
