+++ desktop/apps/auction/components/artwork_rail/ArtworkRail.styl
