+++ desktop/apps/auction/stylesheets/index.styl
@@ -24,12 +24,13 @@
 @require '../components/artwork_browser/header/HeaderMobile'
 @require '../components/artwork_browser/ArtworkBrowser'
 
+@require '../components/artwork_rail/ArtworkRail'
+
 @require '../components/layout/auction_info/AuctionInfo'
 @require '../components/layout/auction_info/Registration'
 @require '../components/layout/Banner'
 @require '../components/layout/Footer'
 @require '../components/layout/active_bids/MyActiveBids'
-@require '../components/layout/artwork_rail/ArtworkRail'
 @require '../components/layout/active_bids/ActiveBidItem'
 
 .auction-body
